requirements.txt
appdirs==1.4.4
attrs==20.3.0
audioread==2.1.9
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
conda==4.9.1
conda-build==3.20.5
conda-package-handling==1.7.0
cryptography==2.9.2
dataclasses==0.6
decorator==4.4.2
dnspython==2.0.0
filelock==3.0.12
future==0.18.2
glob2==0.7
h5py==3.1.0
idna==2.9
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
libarchive-c==2.9
librosa==0.8.0
llvmlite==0.34.0
MarkupSafe==1.1.1
mir-eval==0.6
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
musdb==0.3.1
museval==0.3.1
numba==0.51.2
numpy==1.19.2
olefile==0.46
packaging==20.4
pandas==1.1.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
pkginfo==1.6.0
pooch==1.2.0
prompt-toolkit==3.0.8
psutil==5.7.2
ptyprocess==0.6.0
pyaml==20.4.0
pycosat==0.6.3
pycparser==2.20
Pygments==2.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-etcd==0.4.5
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
resampy==0.2.2
ruamel-yaml==0.15.87
scikit-learn==0.23.2
scipy==1.5.4
simplejson==3.17.2
six==1.14.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
stempeg==0.1.8
threadpoolctl==2.1.0
torch==1.7.0
torchelastic==0.2.1
torchvision==0.8.0
tqdm==4.46.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.8
wcwidth==0.2.5
